[inscon-2] fixed data type of Int16 to Int32

+++ Enterprise/Database/EnterpriseDb/Stored Procedures/crud.LookupGroup_Update.sql
@@ -1,10 +1,11 @@
+
 SET QUOTED_IDENTIFIER ON
 GO
 SET ANSI_NULLS ON
 GO
 
 CREATE PROC [crud].[LookupGroup_Update]
-  @LookupGroupID SMALLINT,
+  @LookupGroupID INT,
   @Name          VARCHAR(75)
 AS
     SET NOCOUNT ON
